(Still Kevin)
Cleanups to allow different kind of pipelines to coexist in the same library and to eliminate various Makefile level build switching.  For example, while the various stages of the Cell port are incomplete, it will be necessary to run some scalar pipelines on the PPE and run some SPU pipelines elsewhere.  This requires multiple pipelining models to be present and multiple kinds of pipelines to be present simultaneously.