(Still Kevin.  I am supposed to be able to check out as me now, so hopefully this will be the last commit from me through Ben's svn.)
(1) Completed the installation of the scalar particle pipelines.  There was
    some minor numerical tuning in center_p and uncenter_p of these in the
    process to make their round-off behavior closer to advance_p.

(2) Fixed a minor flaw in the pipeline API.  Namely, pipelines do not know
    which pipeline dispatcher called them.  However, if pipelines are
    responsible for determining what they responsible for computing (instead of
    being explicitly told by the host), they need to know the number of
    pipelines dispatched in addition to which pipeline they are.  When there
    was there are multiple pipeline dispatchers, pipelines cannot reliable
    ascertain this information.  So, this information is now explicit passed to
    the pipeline in addition to the rank.

Compiles clean and produces bit-level identical results on my desktop.  (I was
somewhat surprised about this given the tuning in center_p and uncetner_p but
as these are only used during initialization and for diagnostics, it is
possible that for a short run, it did not affect any diagnostics being
analyzed.)