Removed the __SPU__ define conditional compilation that appeared.
First, this executes on the PPU side of the interface so it should be
compiled by the PPU compiler, not the SPU compiler.  Second, for technical
reasons, we should have our own defines control any compile time switching
that occurs in a file like this (e.g. a "CELL_SPU" defined in machine
description) rather than trust the various compilers we might use as there
are all sorts of interesting things I am discovering in bringing up the
complete advance_p pipeline.