Minor API tweak to allow the more efficient threading on homogeneous architecture multi-core shared memory situations (e.g. the two cores of the PPU).
Specifically, the threaded pipeline model has two possible dispatch behaviors.
In the first behavior (dispatch_to_host), the host thread is treated as a
pipeline thread and thread.dispatch will execute one of the pipeline workloads
on the host thread.  In the second, the host thread is not treated as a
pipeline thread and the host thread can do other things while pipeline threads
are processing workloads.

The latter model is what the SPU dispatcher must use (e.g. a PPU host thread
is simply incapable of running a SPU pipeline).  The latter model is more
more efficient and easy to program for shared memory multiple processing
(e.g. distributing a job across the two PPUs on the PPE).  It also eliminates
the need for the serial dispatcher in production usage as the serial dispatcher
is equivalent to the thread dispatcher with one thread and dispatch_to_host 
true.)