Fixed a divide-by-zero bug when computing the index of first bulk voxel to process in distribute_voxels.  This bug would trigger when the local voxel mesh was so small that the bulk voxel region used for pipelines contained no voxels (e.g. 1d and probably some 2d simulations).
Note that even in the presence of the divide-by-zero, the number of voxels
assigned to each pipeline was still correct.  Thus, technically, this bug
could not affect computations (i.e. a pipeline would have been told to process
zero voxels starting from an invalid voxel index).  However, some platforms
(presumably Cell) are not forgiving to divide-by-zeros and fault when they
occur.  So the voxel index computation is now well defined with no
divide-by-zeros for any region parameters.

Hat tip to Ben for finding the bug.

(Staggered meshes are fun to efficiently index.)