Some changes to help Ben hook up advance_p_pipeline for Roadrunner performance modeling.
(1) advance_p_pipeline_spu.c is used as highest performance complete (and
thought to be correct) particle pusher.  The old C++ version was moved into
the tmp/advance_p harness directory for use with further testing and the
current header for the C-language baby v4 (v4c_spu.h) needed by
advance_p_pipeline_spu.c was copied into util/v4.  Note that this header is
not robust so I don't recommend taking anything in there as final.

(2) Previously, I was confused on the number of Power cores on the Cell due to
how the simulator presents them.  As a result, I changed how the number of
PPE pipeline threads is set.  Now, the number of PPE pipeline threads is set
via a threads per process "-tpp=" command line argument while the number of
SPU pipelines is hardwired to 8.  Given that there is a single dual threaded
Power core rather than a dual single threaded Power cores on Cell and for
other technical reasons how SPE pipeline dispatch works, the default of
"-tpp=1" is most probably appropriate for most usage (but it can be run-time
configured if not).

(3) Made some minor cleanups to the harness in tmp so I could test them on
Bullock again and to facilitate switching pipeline lines in the future
if advance_p_pipeline_spu_{2,4} become viable.