Cleaned up some problems with revision 154 and fixed a bug in the PPU / SPU advance_p pipeline.  Note:
(1) CELL_PPU_BUILD - Is defined by the build infrastructure and no one else.
It is defined if and only if the target machine code for the source code being
compiled is a Cell PPU.

(2) CELL_SPU_BUILD - Is defined by the build infrastructure and no one else.
It is defined if and only if the target machine code for the source code being
compiled is a Cell SPU.

(3) USE_CELL_SPUS - Is defined by a machine description and no one else.  It
indicates to the source code that the SPU pipeline implementations should be
utilized when present and applicable.

(4) USE_V4_* - Is defined by a particular machine description and no
one else.  It indicates to the source code that V4 pipeline implementations
should be utilized when present and applicable.  The * can be
{PORTABLE,SSE,ALTIVEC,SPU} depending on particular machine description.

(5) HAS_SPU_PIPELINE - Is defined case-by-case in code that provides the
interface to a particular pipeline before including any headers and no one
else.  It indicates to various headers that there is a SPU implementation of
the pipeline.

(6) HAS_V4_PIPELINE - Is defined case-by-case in code that provides the
interface to a particular pipeline before including any headers and no one
else.  It indicates to various headers that there is a V4 implementation of
the pipeline.

All pipelines have a scalar implementation.  The scalar variant is used to
process stragglers and used as a portable fallback solution when neither
V4 nor SPU implementations of a particular pipeline are available.  When
multiple implementations of a pipeline are available, the order of precedence
for bulk pipeline processing is:

(1) Use the SPU implementation.
(2) Use the V4 implementation.
(3) Use the scalar implementation.

As far as I know, the above is sufficient to handle all combinations of
builds including all mixtures of scalar, V4 and SPU pipelines, x86, AltiVec and
general platforms ...  flags like SPU_ADVANCE_P and USE_SPU_PIPELINE are
superfluous given the above.

In short, to build on the Cell running as a pure scalar code on the PPUs,
the machine description needs nothing special.  To build on Cell running as a
scalar code on the PPUs but using the SPU pipelines, define USE_CELL_SPUS
in the machine description.  To build on Cells running as an AltiVec code
on the PPUs without the SPU pipelines, define USE_V4_ALTIVEC.  To build
on Cells running as an AltiVec code on the PPUs with the SPUS, define both
USE_CELL_SPUS and USE_V4_ALTIVEC.

If you have questions or if there are bugs that prevent the code from
building this way, please consult with me before changing the above system
of flags.