I found a reasonable way around IBM's 32-bit cache api dubiousness and I have the harris input deck running successfully with multiple MPI processes and multiple SPUs simultaneously with "reasonable" results (given the floating point rounding horrors of single precision on the SPUs).
Thus, the port should now be fully functionality.  It is worth noting that the
SPU pipeline itself required _no_ modifications through all this.  Thus, the
harness measurements inside and outside the simulator are still valid.
However, please note that the spu pipeline dispatch is remarkable slow for
unknown reasons and this will affect end-to-end benchmarks when running below
~1e7 particles per Cell.  This will have to be addressed in the near future.

The tweaked cache-api caches a segment of external memory no more than 4GB in
size.  The addresses passed by the user are relative effective addresses into
this segment.  Because the window is at most 4GB in size, the relative
effective addresses need only be 32-bits in size.  As such, all of IBM's fancy
tricks based on 32-bit integer SIMD for cache lookup and what not will still
work without modification.

The base of the segment of cached memory is given by "CACHE_BASE_EA" which the
user species when defining the cache.  If CACHE_BASE_EA is not defined, it is
assumed to be zero and the cache-api works _exactly_ like it worked before
(i.e. it only needs to use the 32-bit address based DMA transfers).  If
CACHE_BASE_EA is defined, the cache will use 64-bit effective addresses for the
DMA transers, as, in general, some of the memory in the window may require
64-bit effective addresses.  CACHE_BASE_EA must be 128 byte aligned.

The performance hit of this should be negligible.  It only affects cache misses
and writebacks and amounts to a few extra integer ops for the address
calculation in these cases.

(Yay ... all that experience in my youth with segmented memory on my first home
computer---an IBM PCjr---finally came in useful.)

A patched "cache-4way.h" has been temporarily added to the
src/particle/pipelines_cell directory.  The other Cell SDK cache-api files have
also been added so that the patched cache-4way.h will be included.  These other
files were not modified though.

NOTE: I DO NOT KNOW THE LICENSE FOR THESE FILES.  THUS SO LONG AS THESE FILES
ARE IN THE VPIC SOURCE TREE, VPIC SHOULD NOT BE DISTRIBUTED.  PLEASE BEAT IBM
WITH A RUBBER HOSE UNTIL THEY PATCH THEIR CACHE-API AS DESCRIBED ABOVE AND KEEP
VPIC FREE FROM ANY LICENSE HEADACHES.

I also made some commenting cleanups in standard_cell_64 and
standard_cell_ppe_64.