Some infrastructure tweaks in preparation for installing the spu quad-quad particle advance pipeline.
- The padding necessary to align the interpolators to 128-byte boundaries for
the benefit of the SPUs DMA controller is large enough to accomodate all
the neighbor information for the voxel as well.  Thus, when building
on a spu, the neighbor data is copied into the padding of the interpolator
when loading the interpolator.  This allows the complete elimination of
the neighbor caching in the spu advance_p pipelines, allows neighbor data to be
pulled into the SPU for free with the interpolator cache and allows for large
cache datastructures in the spu local store.  (A big win for warm particles.)

- A few more primitives were added to v4c_spu.h.