Added script to run the tests using the relay.
Modified the relay so that checks pending messages on a counter, rather than
checking every time it polls.  This could cut down some on the latency of
using the relay.  I'll have to test it some to see if it makes a huge
difference.  This change just adds the knob.