Hooked up the collision model stubs.  This involved uncommenting text checked in earlier into advance.cxx and deck_wrapper.cxx.  Also had to add a declaration to vpic.hxx
This change broke all the decks in sample/ so blank stubs were added to the 
ends of each deck.  Updates were made to almost all of the input decks to 
make them compliant with the new source tree.  Presently, all decks compile 
with the exception of sample/test/v4, which dies because of an ambiguous 
reference to all() (this looks like a uint32_t/int conflict).  Also, 
SC08_ONE_TRIBLADE_DEMO doesn't compile; I wasn't sure what to do with 
references to checksums among other things.  

A minor bug was fixed in the volume and surface emitter interface in 
deck_wrapper.cxx (argument must be a pointer to species_t and not a 
species_id). 

Made a slight change to define_material() in the most general case; magnetic 
conductivities should default to zero for dielectric materials like calcite.  

Minor touch-up of formatting in child-langmuir.c