Fixed a bug with placement of the collision stub.  (Apparently, I couldn't read my own comments).
Added a fixme to dump.cxx - we should in principle be able to restart a run with no 
species defined.

Added some defaults to inject_particle() to be backward compatable with earlier 
decks.  

Added a runtime MESSAGE to indicate how many threads per processor are being
used.  (The command-line handler does this silently with defaults; it would be 
preferable for the user to know if this is being parsed properly).  

Fixed an erroneous comment in species_advance.h

Added machine files for lobo and flash64 openmpi builds under the old build system.