Major performance boost to VPIC (commodity platforms).
In short, 10% performance boost to VPIC for particle
dominate simulations with typical particle temperatures
(e.g. cold ion species, and a wpdt = 0.2, lambda = debye
electron species), with greater impact on simulations
with very warm particles.

This isn't completely polished yet (and thus this check-in does
not fully enable the new code being checked in) but it does seem
to work.  To play around with it in the meantime (don't know when
I'll get a chance to polish it), change the Imakefile to compile
new_move_p.cxx instead of move_p.c (and similarly for the autoconf
build infrastructure).

Changes:

(1) V4_SSE was revamped for higher performance on modern
implementations of SSE.  This gave a 7% performance boost to
cold particle benchmark (e.g. my desktop (bullock) is can now
pull 24.4 MPA/s instead of 22.8 MPA/s).  The other V4
implementations need minor tweaks to mirror the revamped SSE
V4 implementation (all existing V4 accelerated code will work
regardless).  It might be possible to get more juice out
of the revamped SSE implementation as I didn't do detailed
performance tuning of it.  Likewise, there are some tricks
that I probably should do to the cold advance pipeline to
allow the compiler greater freedom to optimize it.

(2) The vectorized accelerated version of move_p used on the
SPE was backported to commodity platforms.  We don't have any good
end to end benchmarks that test the performance of move_p but this
implementation is roughly 3X faster than the existing move_p.
Practically, for the harris input deck (which is close to being
a cold particle advance), the particle advance time was
accelerated an additional 6% beyond the 7% accleration above.
This is not fully tuned either, but, even in its present state,
this should made a noticeable impact on things like relativistic
pair plasma input decks.  The new move_p has different but
equivalent roundoff properties as the old move_p (directly
measured).

Enjoy,
Kevin