(1) Installed the option for doing current accumulator accumulation in     advance_p_pipeline_spu with IEEE floating point rounding.
    This option is currently ENABLED.  To disable, open advance_p_pipeline_spu
    and comment the line before the particle advance kernel that says
    "#define USE_DOUBLE_PRECISION_REDUCTION".

    This has been measured to reduce divergence errors in J by over an order of
    magnitude on Cell and likely will also eliminate the divergence error
    spatial patterns Bill and (to a lesser extent) Lin has seen on Roadrunner
    (see the theory note).

    This option is _NOT_ free.  It shaves around 15% off the particle advance
    kernel performance (floating point performance is roughly the same if I
    count the extra operations necessary to get the SPE to round current
    accumulator reduction arithmetic correctly).  However, I don't expect this
    to be an issue end-to-end.  Specifically, divergence cleaning can be run an
    order of magnitude less frequently as a result of this and div cleaning is
    a major time sink on Roadrunner currently.

    Also, something that might be of more general interest, in benchmarking
    this change, I noticed on rr-dev, particle advance performance is
    saturating between 7 and 8 SPE.  Since certain nasty PPE overheads
    current grow with the number of SPE used, it might be worthwhile to
    experiment with using fewer SPEs to accelerate end-to-end performance.

(2) Cleaned up the unit test infrastructure a bit.  Notable changes are the
    unit tests have much better coverage when there are multiple SPE used, the
    tests are much more mathematically rigorous both theoretically and in
    implementation and the tests all behave consistently from the shell
    point of view unit test on success and failure (allowing the tests to
    be more extensively automated).

(3) Backported the unit test harness from VPIC-RZ ("verify") and hacked it up
    to run VPIC's unit tests on Roadrunner in an automated fashion and under
    a wider variety of threading situations.

(4) Cleaned up the command line options for the asds_64 machine (the brutality
    is restored).

Theory note:

Suppose you plan to sum up N numbers which, on average, are 1 (e.g. current
reduction for all particles in a cell of a drifting plasma).  This yields N.

Suppose you are on a machine with truncating floating point arithmetic and you
have the partial sum of the first n numbers S_n.  Then, roughly, the partial
sum of the first n+1 numbers S_{n+1} is on average:

  S_{n+1} = (S_n + 1)( 1 - e/2 )

where e ~ FLT_EPSILON and the 1-e/2 above is due to the fact that the answer on
averge will be e/2 fractionally too small when using truncating arithmetic.
Solving this difference equation yields (with boundary conditions S_0 = 0 and
S_1 = 1-e/2 for simplicity):

  S_N = (2/e - 1)( 1 - (1-e/2)^N )

keeping all terms up to first order in e yields the bias in the reduction
due to using truncating arithmetic:

  dS_N ~ -(e/4)( N^2 + N ) ~ -(e/4) N^2

Imagine if this sum is broken into two parts that are performed independently
(e.g. one thread reduces half the particles in a cell, a second thread reduces
the other half and the two halfs are then summed).  Then the overall bias
becomes:

  2 dS_{N/2} ~ -(e/8) N^2.

which is _different_ than the bias when the sum is performed all at once.

Because particles are sorted, the cells for particle current reduction are
split have a recognizable spatial pattern.  (E.g. For a uniform particle
distribution in large cubic domain, particle currents in cells near the domain
boundary and near x-y planes that divide the local domain n_thread equal
chunks would be reduced by two or more threads while particle currents in
other cells would be reduced by one thread.)

IEEE compliance is a good idea, _ESPECIALLY_ in single precision.

In short, oh no.