(1) Synchronized Cell accelerated code with r413.
(2) Fixed clear / reduce accumulators.  It turns out that reduce accumulators
was the problematic one but it was problematic in a tricky way not likely
to be caught by the unit tests (basically, DMA transfer implementation in
reduce_accumulators was broken in such a way that it would usually pass on the
types of toy simulations used in unit tests but was badly and
non-deterministically behaved in production usage).

(3) Tuned up the asds_64 machine compile options.

(4) Tested this heavily under many exotic configurations (with new
move_p, with the old move_p, with various SPE accelerated pipelines
enabled or not, with overlays LD setup in different ways, with all
manners of tpp and spp settings, unit tests, benchmarks, various sample
decks).  Seems to work fine.