Finished debugging SPE particle advance.  It turns out that what I checked in previously was fully functional.  However, the restriction on the number of voxels that can processed by each subsort was slightly too large.  And it turns out that Cell and friends go out of their way to make it hard to tell when you've exhausted your local store.
Also recombined the files that were separated and did some polishing of the
final implementation.

The new particle sorting should be many times faster than the old particle
sorting (particularly on large and/or 3d simulations).

This check in should not change any baselines.

Still to do: Reenable the old in-place sorting option