Based on Bill's runs, enabled the high performance mover on x86 and disabled it on Cell.
Careful examination of the Cell V4 implementations shows that Altivec
doesn't have a vector division instruction and the implementation of
division used in V4 would not be completely IEEE compliant (it is the
recommended implementation from Apple's developer web site).  If
there isnt a compiler bug, then I _strongly_ suspect this is the
cause of the problem.  At some point in the near future, I'll change
how Altivec does division and see if that fixes Bill's problem.