This will be the first of _many_ changes to VPIC (this is bring broken up to keep the change logs managable).  This battery of checkins radically improve the underlying VPIC infrastructure and fix almost every known architectural issue with VPIC.
For example:

- You can do things like checkpt and restart from a modified executable.
- You can have users add exotic functionality to VPIC and have it
  automagically checkpt correctly.

- Custom boundary conditions now know the simulation time.

- You no longer need to widen the interface of the boundary handler everytime
  you add a function.

- You can use non standard (and even user specified) custom field solvers and
  have the code checkpoint

- You can have arbitrary large parameter sets for custom emitters and
  boundary conditions.
- Emitters can emit multiple species if so desired.

- The rank and size of process is always visible.

- ...

It also fixes several years of accumulated bit rot / cruft (i.e. the decay
of code with time) and scrum (i.e. ugly hacks added to the code to make it work
on short notice.  For example:

- VPIC's objects now have well defined semanatics (needed for a checkpting).

- Command line parsing is rational.

- The messing layer API is somewhat sensible (could be better but is no longer
  the house of shame it was previously).

- The interface between mp and relay has lost two unnecessary layers of
  abstraction.

- Tons of cargo cult code has been been refactored and reduced size 
  greatly.

A few minor bugs were fixed as well:

- uncenter_p was actually doing centering (only affects simulation 
  initialization and then in ways unlikely to be visible to any existing
  input args).
- ...

The checkins have been heavily tested on my home system but these checkins:

- Change some of the command line and input deck APIs.  As such, existing
  input decks and job launch scripts will break.  I have updated all the
  sample input decks to compile under the new API (and have heavily tested
  "harris").  It is straightforward to update input decks to the new APIs
  (the input decks generally get much cleaner as a result).

- Will break the new build system (haven't updated it) and thus will not
  compile on Roadrunner.

After all this gets submitted, I'll check in the new collision module.
(In writing the collision model, I had to change the species.h data
structures.  As such, I had to modify almost every piece of code in VPIC which
triggered the above refactoring.)

Each checkpt will document the changes revelant to that check in.

About this check-in:

This is the new checkpt service.  When you create some state that needs to
be checkpointed, register how to serialize and restore it with this service.
This service provides a large number of helpers to assist with writing
individual checkpt/restore functions (including checkpointing function
pointers reliably across different builds of the code).

Note that this checkpoint service is _blind_.  That is, it knows nothing
about VPIC and, as such, can be used to checkpt/restore anything.  As such,
checkpt is fully decentralized and if, for example, you want to append
some state in your input deck to a checkpt file, go right ahead, right
from your input deck.  It will work, without need to touch anything inside
VPIC.  This allows, among other things, the amazing central planning horror
that was "dump_restart" to be eliminated _entirely_ from VPIC.

See checkpt.h for extensive documentation and the rest of the code for
examples of various individual checkpt/restore functions work and how
to write a checkpt and restore from a checkpt (both in from a developer
and user point of view).

For Ben: Feel free to make a policy abstract like the refactored mp policies
so that checkpt_io can use your FileIO primitives.