Changed basic species data types to accomodate the collision package. - Particle charge => # physical particles represented - Species charge/mass ratio => separate physical particle charge and physical   particle mass
A species now knows the grid it is on.  This allows a great many APIs and
what not to be simplified and deuglified.  For example, the species partition
can be created immediately (rather than deferred to a hack in first sort_p
call)

Hooks up the species_t to the checkpt service.

Minor algorithm improvement to num_species (now that species_list have
well defined semantics).

Eliminated the silly invalid_species_id (just pollutes the namespace ...
worthless).

Minor API and code consistency and API changes.