Major revamp of the emitter API (a good example of the power that the lower level changes give you).
Latent physics bug fixed:  The normal distribution for thermal emission was not 
being picked with the correct distribution randomly (the net flux and zero
temperature limits were correct though).  This was a dup of the older bug in
maxwellian_reflux and is not thought to have affected any simulations.

Emitters are very flexible now.  There is no restriction on the number of
parameters used to set up an emitter, the amount of state held by an emitter,
the types of state that can be checkpointed by an emitter (e.g. you can
have emitters with arbitrary object and function pointer dependencies), what
an emitter function can do ... in fact the begin_particle_injection section
of the input deck is now superfluous given the new API (but I'll keep it around
this input deck section for user convenience).

Each emitter type provides the user a custom constructor.  The caller than
sizes the emitter to accomodate the geometry that does emission and then
populates the emitter with that geometry.  (From an input deck point of view,
the last two steps are done in the simultaneously.)

Emitters lost their name (the rule right now is that structures have
human readable names if they are likely to produce errors in contexts where
the human readable name could be helpful in debugging / disambiguating the
which object generated the error).

Ivory and ccube emitters were absorbed in the child-langmuir emitter.

Minor code formatting cleanups.