Installed hard sphere collisions.  Non-relativistic particle-particle and particle-fluid hard sphere collisions are supported.  The particle-fluid model allows the fluid to be warm and drifting.
The support uses the unary and binary collisions models and is extensively
documented (complete with derivations of optimized rate constant function
calculations and the optimized microscopic physics calculations).

Since, theoretically, all binary elastic collision processes are the same
(the only differences are cross section and how the scattering angle is
picked in the reduced mass system), this should serve as a robust template
for _all_ fluid-particle and particle-particle collision processes in the
none relativistic limit.

P.S. Since SVN is still likely broken, I'll attach a figure showing
graphically a minor approximation made in the rate constant calculation
in the warm fluid case to dramatically speed it up.  The derivation of
this approximation is in the source code.