Created a collision model benchmark and put each collision model through its paces.  As it stands, they all pass the smell test (the sampling methods, the approximations for the fluid background and so forth yielding the correct collision rates for physical particles).
I think the hard sphere collision operators have a bug in their momentum
transfer calculations though as it is heating benchmark system very fast.
Large angle coulomb, which only differs in the momentum transfer
calculation from hard sphere, is not heating the system much.  In any case,
for the purposes of the L2 milestone, these will give a very accurate idea
for performance.