Updated make_version script.
This will use commits from the most current tag to create a version number.
After this check-in, I will add a tag "2.0", so that subsequent commits will
generate version info like 2.0.1, 2.0.3.  These are purely for ease-of-use and
have no effect on the repository.  If anyone objects to "2.0", we can easily
create some other tag.  The 2.0 tag will reference the version that is
"complete" to this point, i.e., includes test framework, bug fixes, OSX build
fixes and is the current build that Brian, et al. are using.  Tag 531 marks the
initial version that was migrated from the SVN repository.  There is also a tag
for version 407.

To revert to a tag, clone the repository and then "checkout" the tag:

% git clone git@ccs-dev-fe:vpic.git
% git checkout <tag name>

To go back to the master:
% git checkout master