Add initial support for V8, the new vector class to support SIMD vectors of length 8.  The strategy for accomplishing this work is divided into three phases with the possibility of an optional fourth phase.  The first phase is to add support for a V8 class which is actually a clone of the V4 class with the names changed.  But the initial V8 class really implements SIMD vectors of length 4.  This will help to verify that V8 support has been added to all the files where it is required and also to make sure that the required build support has been added.
The second phase will be to convert the V8 portable implementation into a
properly functioning implementation with SIMD vectors of length 8.

The third phase will be to copy the V8 portable implementation to a V8 AVX2
implementation and then incrementally convert the portable implementation to
an AVX2 implementation.

The optional fourth phase would be to then study the similarities between the
V4 and V8 implementations in order to determine if a better implementation
could be designed using either generic or object oriented C++ programming
techniques.