Comment out CELL specific code for now in anticipation that these remaining references to CELL will be removed like the rest of the CELL stuff previously.
Add the ability to reverse the order of boot sequence for the communication
layer and the pipelines.  This is accomplished by defining the CPP variable
VPIC_SWAP_MPI_PTHREAD_INIT at compile time.  The default is the original
order where the pthreads are created before MPI is initialized.  The new
reversed order is untested at this time.