Documentation of VPIC CMake build variables (#52)
* Add configurable and documented build scripts for building VPIC on LANL ATS-1 and CTS-1 machines.
Document how to use these two scripts.

* Additional updates to documentation.

* Update compiler option documentation to make more accurate.

* Reorder options a bit.

* For the lanl-ats1 script, make sure that the Cray programming environment
starts out as the Cray default of PrgEnv-intel. This change checks for the
case where the user has modified their module environment and swaps it back
to the case assumed by the build script.

* Fix issues and errors in use of float literals introduced in a previous commit.

* Add CMake support for configuring a VPIC build with the legacy particle sort
implementation. Add build script support for a few more CMake variables that
were missing and should be availble to users of the build scripts.

* Separate lanl-ats1 script into two separate scripts, one for Haswell nodes and one for KNL nodes.

* Reorder build configuration options to something that seems more sensible.

* Add more documentation about various available CMake configuration variables.