support vpic config using a cmake cache file
The current set of config scripts in arch configure vpic by running
cmake with the config specified on the cmake command line using cmake
cache variables (i.e. the "-D" flag).  The lanl scripts also force
the module config and run make.  There are two problems with this.
First, users no longer have access to the cmake command line (e.g.
so that they can specify the installation prefix for "make install"
with -DCMAKE_INSTALL_PREFIX=...).  Second, the config processes on
lanl systems differs from non-lanl systems.

This patch adds a new optional way to configure vpic with cmake
that works around these issues.   When an arch script is invoked
with the optional arg "initcache" it generates a file called
"vpic-init-cache.cmake" and stops.  The "vpic-init-cache.cmake"
file contains all the cache variables that the arch script would
have normally put on the command line (with the "-D" flag).

This allows the user (or automated build script) to run cmake
themselves with additional flags by using the "-C" flag to load
the "vpic-init-cache.cmake" file.

For example, to build a test vpic install in /tmp/try one can do:

    % ls
    % ../arch/gcc/v4-sse initcache
    ./vpic-init-cache.cmake created.
    Now you can run "cmake -C ./vpic-init-cache.cmake src_dir" to configure.
    %
    % ls
vpic-init-cache.cmake
    %
    % cat vpic-init-cache.cmake
    set(CMAKE_BUILD_TYPE "Release" CACHE STRING "Initial cache" FORCE)
    set(ENABLE_INTEGRATED_TESTS "ON" CACHE BOOL "Initial cache" FORCE)
    set(USE_V4_SSE "ON" CACHE BOOL "Initial cache" FORCE)
    set(CMAKE_C_FLAGS "-rdynamic -fno-strict-aliasing" CACHE STRING "Initial cache" FORCE)
    set(CMAKE_CXX_FLAGS "-rdynamic -fno-strict-aliasing" CACHE STRING "Initial cache" FORCE)
    %
    % cmake -C ./vpic-init-cache.cmake -DCMAKE_INSTALL_PREFIX=/tmp/try ..
    loading initial cache file ./vpic-init-cache.cmake
    -- The C compiler identification is GNU 7.4.0
    -- The CXX compiler identification is GNU 7.4.0
    -- Check for working C compiler: /usr/bin/cc
    -- Check for working C compiler: /usr/bin/cc -- works
    -- Detecting C compiler ABI info
    -- Detecting C compiler ABI info - done
    -- Detecting C compile features
    -- Detecting C compile features - done
    -- Check for working CXX compiler: /usr/bin/c++
    -- Check for working CXX compiler: /usr/bin/c++ -- works
    -- Detecting CXX compiler ABI info
    -- Detecting CXX compiler ABI info - done
    -- Detecting CXX compile features
    -- Detecting CXX compile features - done
    -- Found MPI_C: /usr/lib/x86_64-linux-gnu/libmpich.so (found version "3.1")
    -- Found MPI_CXX: /usr/lib/x86_64-linux-gnu/libmpichcxx.so (found version "3.1")
    -- Found MPI: TRUE (found version "3.1")
    -- Looking for pthread.h
    -- Looking for pthread.h - found
    -- Looking for pthread_create
    -- Looking for pthread_create - not found
    -- Looking for pthread_create in pthreads
    -- Looking for pthread_create in pthreads - not found
    -- Looking for pthread_create in pthread
    -- Looking for pthread_create in pthread - found
    -- Found Threads: TRUE
    -- Configuring done
    -- Generating done
    -- Build files have been written to: /proj/TableFS/data/chuck/src/lanl/devel/vpic/b
    %

then a "make install" would build and install vpic in CMAKE_INSTALL_PREFIX
("/tmp/try" in the example above).

To use this with LANL-style builds, we decouple the module configuration
from vpic configuration.   We assume LANL users will configure all their
modules the way they want them prior to trying to configure vpic.  We do
not change the module config.   Instead, we look into the environment to
determine the current module configuration and use that.   To support
this mode of operation we introduce a new script "arch/lanl-current-mods"
that combines ATS1 and CTS1 configs into a single script named
"lanl-current-mods" ... this is used in the same way as noted above,
but internally it includes extra code to examine and sanity check the
currently loaded set of modules.  Here is an example from GR:

    [ccranor@gr-fe2 b]$ ../arch/lanl-current-mods initcache
    ERROR: No modules loaded.  A compiler, MPI, and cmake are required.
    [ccranor@gr-fe2 b]$
    [ccranor@gr-fe2 b]$ module load gcc openmpi cmake
    [ccranor@gr-fe2 b]$
    [ccranor@gr-fe2 b]$ ../arch/lanl-current-mods initcache
    Compiler loaded: GNU version 7.4.0
    MPI loaded: OMPI version 2.1.2
    CPU target: broadwell

    Currently Loaded Modules:
      1) gcc/7.4.0   2) openmpi/2.1.2   3) cmake/3.14.6

    ./vpic-init-cache.cmake created.
    Now you can run "cmake -C ./vpic-init-cache.cmake src_dir" to configure.
    [ccranor@gr-fe2 b]$
    [ccranor@gr-fe2 b]$ cat vpic-init-cache.cmake
    set(CMAKE_BUILD_TYPE "Release" CACHE STRING "Initial cache" FORCE)
    set(ENABLE_INTEGRATED_TESTS "OFF" CACHE BOOL "Initial cache" FORCE)
    set(ENABLE_UNIT_TESTS "OFF" CACHE BOOL "Initial cache" FORCE)
    set(ENABLE_OPENSSL "OFF" CACHE BOOL "Initial cache" FORCE)
    set(DISABLE_DYNAMIC_RESIZING "OFF" CACHE BOOL "Initial cache" FORCE)
    set(SET_MIN_NUM_PARTICLES "128" CACHE STRING "Initial cache" FORCE)
    set(USE_LEGACY_SORT "ON" CACHE BOOL "Initial cache" FORCE)
    set(USE_V4_PORTABLE "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_V4_SSE "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_V4_AVX "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_V4_AVX2 "ON" CACHE BOOL "Initial cache" FORCE)
    set(USE_V8_PORTABLE "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_V8_AVX "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_V8_AVX2 "ON" CACHE BOOL "Initial cache" FORCE)
    set(USE_V16_PORTABLE "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_V16_AVX512 "OFF" CACHE BOOL "Initial cache" FORCE)
    set(VPIC_PRINT_MORE_DIGITS "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_OPENMP "OFF" CACHE BOOL "Initial cache" FORCE)
    set(USE_PTHREADS "ON" CACHE BOOL "Initial cache" FORCE)
    set(BUILD_SHARED_LIBS "OFF" CACHE BOOL "Initial cache" FORCE)
    set(CMAKE_C_COMPILER "mpicc" CACHE STRING "Initial cache" FORCE)
    set(CMAKE_CXX_COMPILER "mpicxx" CACHE STRING "Initial cache" FORCE)
    set(CMAKE_C_FLAGS "-g -O2 -ffast-math -fno-unsafe-math-optimizations -fomit-frame-pointer -fno-strict-aliasing -Winline -rdynamic -march=broadwell" CACHE STRING "Initial cache" FORCE)
    set(CMAKE_CXX_FLAGS "-g -O2 -ffast-math -fno-unsafe-math-optimizations -fomit-frame-pointer -fno-strict-aliasing -Winline -rdynamic -march=broadwell" CACHE STRING "Initial cache" FORCE)
    [ccranor@gr-fe2 b]$

Now you can simply do:

    % cmake -C ./vpic-init-cache.cmake -DCMAKE_INSTALL_PREFIX=/tmp/try ..

in the same way as the non-lanl example above and then do "make install"
to build.

To make this work, we introduce additional shell functions in
arch/cmake-fns.sh (to create and append to the "./vpic-init-cache.cmake" file)
and arch/module-fns.sh (to examine + sanity check the currently loaded
modules and use that into to set VCOM, VMPI, and VCPU).

The new "arch/lanl-current-mods" script is a combined version of
lanl-ats1-hsw, lanl-ats1-knl, and lanl-cts1 adapted to work with
arch/module-fns.sh.

The arch/gcc/* files were also adapted to use "arch/cmake-fns.sh" and
support the "initcache" arg.